refactor: simplify console server — remove dead code, merge serialization, fix safety bugs
Systematically addresses 13 issues from console_simplify_review.md:

- Remove dead code: PaginatedResponse, serialize_step_delta_payload, unused
  cors_origins config, task_projection.py, domain/remote_workspace.py
- Consolidate serialization: merge serialization/{agent,observability,scheduler}.py
  into response_serialization.py, delete serialization/ directory
- Remove duplicate endpoint: /api/scheduler/chat/{agent_id} (identical to
  /api/chat/{agent_id}), move cancel route to chat router
- Fix data safety: remove DROP TABLE IF EXISTS feishu_session that wiped
  session data on every SQLite reconnect
- Fix SessionManager lifecycle: factory no longer creates a throwaway manager;
  commands are built after BaseChannelService creates the canonical one
- Inline ConsoleToolBuildContext single-field wrapper
- Fix weak types: SessionAggregate (object → Run/datetime), MongoDB store
  (object → AsyncIOMotorCollection)
- Extract truncate_for_log as shared module-level function
- Merge session_aggregate_to_chat_summary to reuse summary_data
- Add perf TODO annotations on in-memory pagination

Net: -670 lines, 149 tests pass, ruff clean.
Made-with: Cursor
