--- console/server/serialization/__init__.py
@@ -1,1 +1,0 @@
-"""Transport serialization helpers for console API responses."""
